fix building errors
